增加了时间检查记录文档

+++ CSIM/Sparse_HLS/Sparse_HLS/src/sparse.cpp
@@ -11,7 +11,7 @@ void sparse(
 	t_AXI_DataType *inputs,
 	t_AXI_DataType *outputs,
 	int quant_shift,
-	int16_t quant_mul,
+	int quant_mul,
 	bool &sparse_flag)
 {
 // 硬件绑定
